Move linalg rewrites, delete sandbox
Moving caused a circular dependency with tensor.blas. It seems most
linalg rewrites are in the stablize set, so should run before the blas
specializers anyway, so these checks were removed.

This also deleted the unused `spectral_radius_bound` and dummy
`Minimal(Op)`.
