--- pytensor/sandbox/linalg/__init__.py
@@ -1,1 +1,0 @@
-from pytensor.sandbox.linalg.ops import spectral_radius_bound
