+++ pytensor/tensor/rewriting/linalg.py
